chore: rename suggested to suggested-apps
