+++ frontend/src/assets/suggested-apps/alby.png
