+++ frontend/src/assets/suggested-apps/amethyst.png
