+++ frontend/src/assets/suggested-apps/bitcoin-connect.png
