+++ frontend/src/assets/suggested-apps/kiwi.png
