+++ frontend/src/assets/suggested-apps/nostrudel.png
