+++ frontend/src/assets/suggested-apps/primal.png
