+++ frontend/src/assets/suggested-apps/snort.png
