+++ frontend/src/assets/suggested-apps/wavelake.png
