+++ frontend/src/assets/suggested-apps/zap-stream.png
