+++ frontend/src/assets/suggested-apps/zapple-pay.png
