+++ frontend/src/assets/suggested-apps/zappy-bird.png
